+++ pom.xml
@@ -55,7 +55,6 @@
 			<artifactId>tx-control-provider-jpa-local</artifactId>
 			<version>0.0.2</version>
 		</dependency>
-
 		<dependency>
 			<groupId>org.osgi</groupId>
 			<artifactId>org.osgi.service.component.annotations</artifactId>
@@ -119,50 +118,50 @@
 	 	<dependency>
 			<groupId>com.fasterxml.jackson.jaxrs</groupId>
 			<artifactId>jackson-jaxrs-json-provider</artifactId>
-		  	<version>2.8.5</version>
+			<version>2.8.5</version>
 		</dependency>
 		<dependency>
 			<groupId>com.fasterxml.jackson.core</groupId>
-    		<artifactId>jackson-core</artifactId>
-    		<version>2.8.5</version>
+			<artifactId>jackson-core</artifactId>
+			<version>2.8.5</version>
 		</dependency>
 		<dependency>
 			<groupId>com.fasterxml.jackson.jaxrs</groupId>
-    		<artifactId>jackson-jaxrs-base</artifactId>
-    		<version>2.8.5</version>
+			<artifactId>jackson-jaxrs-base</artifactId>
+			<version>2.8.5</version>
 		</dependency>
 		<dependency>
 			<groupId>com.fasterxml.jackson.core</groupId>
-    		<artifactId>jackson-databind</artifactId>
-    		<version>2.8.5</version>
+			<artifactId>jackson-databind</artifactId>
+			<version>2.8.5</version>
 		</dependency>
-	 	<dependency>
+		<dependency>
 			<groupId>com.fasterxml.jackson.core</groupId>
 			<artifactId>jackson-annotations</artifactId>
 			<version>2.8.5</version> 
-		</dependency>       
-        <dependency>
+		</dependency>
+		<dependency>
 			<groupId>org.osgi</groupId>
 			<artifactId>osgi.cmpn</artifactId>
 			<version>6.0.0</version>
 			<scope>provided</scope>
-		</dependency>     
-        <dependency>
+		</dependency>
+		<dependency>
 			<groupId>javax.servlet</groupId>
 			<artifactId>javax.servlet-api</artifactId>
 			<version>3.1.0</version>
-		</dependency>       
-        <dependency>
+		</dependency>
+		<dependency>
 			<groupId>org.glassfish.jersey.core</groupId>
 			<artifactId>jersey-server</artifactId>
 			<version>2.25</version>
-        </dependency>      
-        <dependency>
+		</dependency>
+		<dependency>
 			<groupId>org.glassfish.jersey.containers</groupId>
 			<artifactId>jersey-container-servlet-core</artifactId>
 			<version>2.25</version>
-		</dependency>   
-        <dependency>
+		</dependency>
+		<dependency>
 			<groupId>javax.ws.rs</groupId>
 			<artifactId>javax.ws.rs-api</artifactId>
 			<version>2.0.1</version>
@@ -243,8 +242,7 @@
 			<version>2.5.0</version>
 			<scope>test</scope>
 		</dependency>
-	</dependencies>
-	
+	</dependencies>	
 
 	<build>
 		<plugins>
